Improve column add performance
Add to the table a map of columns by name. This improves
column retrieval by name and duplicated column checks,
specially for tables with lots of columns.

As the column interface allows for column renaming, it was
necessary to store a reference to the table inside the column,
so it is possible to notify the table when a column has its
name changed, and then the table can update its internal map.

This solution is not ideal as we have internal only methods in
both table and column public interfaces.